Добавляет обработчик ESC и обработку отсутствия точек маршрута
