+++ src/components/no-events/no-events-tmpl.js
@@ -1,0 +1,5 @@
+export const createNoEventsTemplate = () => {
+  return (
+    `<p class="trip-events__msg">Click New Event to create your first point</p>`
+  );
+};
