+++ src/components/no-events/no-events.js
@@ -1,0 +1,24 @@
+import {createElement} from '../../utils';
+import {createNoEventsTemplate} from '@Components/no-events/no-events-tmpl';
+
+export default class NoEvents {
+  constructor() {
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createNoEventsTemplate(this._cardData);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
