+++ src/components/trip-controls/trip-controls.js
@@ -1,17 +1,6 @@
-import {createElement} from "../utils.js";
+import {createElement} from '../../utils';
+import {createTripControlsTemplate} from '@Components/trip-controls/trip-controls-tmpl';
 
-const createTripControlsTemplate = (names) => {
-  return (
-    `<nav class="trip-controls__trip-tabs  trip-tabs">
-    <h2 class="visually-hidden">Switch trip view</h2>
-    ${names.map((name) => {
-      return (`
-      <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">${name}</a>
-    `);
-    }).join(``)}
-    </nav>`
-  );
-};
 
 export default class TripControls {
   constructor(names) {
