+++ src/components/trip-day/trip-day-tmpl.js
@@ -1,6 +1,4 @@
-import {createElement} from "../utils.js";
-
-const createTripDayTemplate = (date, dayNumber) => {
+export const createTripDayTemplate = (date, dayNumber) => {
   const currentDate = new Date(date);
   const currentYear = currentDate.getFullYear();
   const currentMonth = currentDate.getMonth();
@@ -13,27 +11,3 @@ const createTripDayTemplate = (date, dayNumber) => {
         <ul class="trip-events__list"></ul>
       </li>`);
 };
-
-export default class TripDay {
-  constructor(date, dayNumber) {
-    this._date = date;
-    this._dayNumber = dayNumber;
-    this._element = null;
-  }
-
-  getTemplate() {
-    return createTripDayTemplate(this._date, this._dayNumber);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-}
