+++ src/components/trip-day/trip-day.js
@@ -1,0 +1,26 @@
+import {createElement} from '../../utils';
+import {createTripDayTemplate} from '@Components/trip-day/trip-day-tmpl';
+
+export default class TripDay {
+  constructor(date, dayNumber) {
+    this._date = date;
+    this._dayNumber = dayNumber;
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createTripDayTemplate(this._date, this._dayNumber);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
