+++ src/components/trip-info/trip-info-tmpl.js
@@ -1,0 +1,6 @@
+export const createTripInfoTemplate = () => {
+  return (
+    `<section class="trip-main__trip-info  trip-info">
+    </section>`
+  );
+};
