+++ src/components/trip-list/trip-list-tmpl.js
@@ -1,0 +1,3 @@
+export const createTripListTemplate = () => {
+  return (`<ul class="trip-days"></ul>`);
+};
